requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-webpack-loader==0.6.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
mysqlclient==1.4.2.post1
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
